backend oldali user authentication és authorization logika kialakítása

+++ webapp/backend/app/Http/Controllers/UserController.php
@@ -7,18 +7,21 @@ use App\Http\Resources\UserResource;
 use App\Http\Resources\UserWithRentalResource;
 use App\Models\User;
 use Illuminate\Http\Resources\Json\JsonResource;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
     public function index(): JsonResource
     {
+        Gate::authorize("viewAny", User::class);
         $users = User::with('person')->get();
         return UserResource::collection($users);
     }
 
     public function store(StoreUserRequest $request)
     {
+        Gate::allows("create", User::class);
         $data = $request->validated();
         $data['pin'] = Hash::make($data['pin']);
         $user = User::Create($data);
@@ -27,12 +30,14 @@ class UserController extends Controller
 
     public function show(User $user): JsonResource
     {
-        $user->load(['cars.fleet']);
+        $user->load(['cars.fleet', 'person']);
+        Gate::authorize("view", $user);
         return new UserWithRentalResource($user);
     }
 
     public function update(StoreUserRequest $request, User $user)
     {
+        Gate::authorize("update", User::class);
         $data = $request->validated();
         $user->update($data);
         return new UserResource($user);
@@ -40,6 +45,7 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        Gate::authorize("delete", $user);
         return ($user->delete()) ? response()->noContent() : abort(500);
     }
 }
